--- src/app/_component/PollsList.tsx
@@ -1,12 +1,0 @@
-"use client";
-import React from "react";
-import useGetPolls from "../hooks/useGetPolls";
-
-type Props = {};
-
-const PollsList = (props: Props) => {
-  const {} = useGetPolls();
-  return <div>PollsList</div>;
-};
-
-export default PollsList;
